+++ streaming/stream_transfer_task.cc
@@ -49,6 +49,7 @@
 #include "message/messaging_service.hh"
 #include "range.hh"
 #include "dht/i_partitioner.hh"
+#include "dht/sharder.hh"
 #include "service/priority_manager.hh"
 #include <boost/range/irange.hpp>
 #include "service/storage_service.hh"
